+++ data-access-service-axon/src/main/java/uk/gov/justice/laa/dstew/access/command/application/priorauthority/DisbursementInformation.java
@@ -1,5 +1,6 @@
 package uk.gov.justice.laa.dstew.access.command.application.priorauthority;
 
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -11,5 +12,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class DisbursementInformation {
-  private Double newAmount;
+  private BigDecimal newAmount;
 }
